pyproject.toml
[tool.poetry]
name = "enigma"
version = "0.1.0"
description = ""
authors = ["Vladimir Chichkin <vovacickin3386@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.10.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
